Namespace oauth under carto to avoid gem conflicts
